The first real commit of the project.
This commit adds " timer_sleep (ticks) " functionality, the implementation
of which is described in part A of the DESIGNDOC (to be committed in the next
commit. just wanted to commit the source first to avoid losing importanbt
data).

WARNING: this code has "not" been tested, and will probably break the program.
This timer_sleep(ticks) commit is just to save the raw un-debugged source
code changes for the first part, which I'm fairly confident in.